+++ shared-helpers/.jest/setup-tests.js
@@ -1,6 +1,9 @@
 // Future home of additional Jest config
 import "@testing-library/jest-dom/extend-expect"
 
+import { addTranslation } from "@bloom-housing/ui-components"
+import generalTranslations from "@bloom-housing/ui-components/src/locales/general.json"
+
 // see: https://jestjs.io/docs/en/manual-mocks#mocking-methods-which-are-not-implemented-in-jsdom
 window.matchMedia = jest.fn().mockImplementation((query) => {
   return {
@@ -14,3 +17,5 @@ window.matchMedia = jest.fn().mockImplementation((query) => {
     dispatchEvent: jest.fn(),
   }
 })
+
+addTranslation(generalTranslations)