+++ sites/public/pages/_app.tsx
@@ -9,8 +9,8 @@ import {
   AuthProvider,
   ConfigProvider,
   LoggedInUserIdleTimeout,
-  blankApplication,
 } from "@bloom-housing/ui-components"
+import { blankApplication } from "@bloom-housing/shared-helpers"
 import { headScript, bodyTopTag, pageChangeHandler } from "../src/customScripts"
 import { AppSubmissionContext } from "../lib/AppSubmissionContext"
 import ApplicationConductor, { loadSavedListing } from "../lib/ApplicationConductor"
@@ -21,7 +21,7 @@ function BloomApp({ Component, router, pageProps }: AppProps) {
   const { locale } = router
   //  const initialized = useState(true)
   const [application, setApplication] = useState(() => {
-    return blankApplication()
+    return { ...blankApplication }
   })
   const [savedListing, setSavedListing] = useState(() => {
     return loadSavedListing()