+++ sites/public/pages/applications/contact/alternate-contact-type.tsx
@@ -14,7 +14,7 @@ import {
   ProgressNav,
   t,
 } from "@bloom-housing/ui-components"
-import { altContactRelationshipKeys } from "@bloom-housing/shared-helpers"
+import { altContactRelationshipKeys, OnClientSide } from "@bloom-housing/shared-helpers"
 import FormsLayout from "../../../layouts/forms"
 import { useForm } from "react-hook-form"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
@@ -50,6 +50,7 @@ const ApplicationAlternateContactType = () => {
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}
           labels={conductor.config.sections.map((label) => t(`t.${label}`))}
+          mounted={OnClientSide()}
         />
       </FormCard>
       <FormCard>