+++ sites/public/pages/applications/household/members-info.tsx
@@ -16,6 +16,7 @@ import FormsLayout from "../../../layouts/forms"
 import { useForm } from "react-hook-form"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
+import { OnClientSide } from "@bloom-housing/shared-helpers"
 
 const ApplicationMembersInfo = () => {
   const { conductor, application, listing } = useFormConductor("householdMemberInfo")
@@ -40,9 +41,9 @@ const ApplicationMembersInfo = () => {
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}
           labels={conductor.config.sections.map((label) => t(`t.${label}`))}
+          mounted={OnClientSide()}
         />
       </FormCard>
-
       <FormCard>
         <FormBackLink
           url={conductor.determinePreviousUrl()}